refactor: colocate tests with source files in cli and chat apps
Move test files from separate `tests/` directories to sit alongside the
source code they test, matching the pattern already used in the API app.

- cli: 15 test files from `tests/unit/` → `src/`
- chat: 9 browser test files from `tests/ui/` → `src/`
- chat: support files (fixtures, page objects, MSW) → `tests/support/`
- Update relative imports, vitest configs, and CLAUDE.md docs
