+++ apps/chat/src/components/chat.render.test.tsx
@@ -1,5 +1,5 @@
 import { describe } from "vitest";
-import { test } from "./chat-page-fixture";
+import { test } from "../../tests/support/chat-page-fixture";
 
 describe("Chat UI - render", () => {
 	test("shows the ready state before a conversation starts", async ({
