+++ apps/chat/src/components/progress-thread.test.tsx
