+++ apps/chat/tests/support/chat-contract.ts
