+++ apps/chat/tests/support/chat-page-fixture.tsx
