+++ apps/cli/src/auth/lib/device-flow.test.ts
@@ -6,8 +6,8 @@ import {
 	requestDeviceAuthorisation,
 } from "@/auth/lib/device-flow";
 import { AuthDeviceFlowError } from "@/errors";
-import { createUnsignedJwt } from "../../helpers/jwt";
-import { mswServer } from "../../setup/msw";
+import { createUnsignedJwt } from "../../../tests/helpers/jwt";
+import { mswServer } from "../../../tests/setup/msw";
 
 const issuer = "http://localhost:8180/realms/meridian";
 const authDeviceUrl = `${issuer}/protocol/openid-connect/auth/device`;
