+++ apps/cli/src/auth/session.test.ts
