+++ apps/cli/src/catalogue/registry.test.ts
