+++ apps/cli/src/cli.test.ts
@@ -1,10 +1,10 @@
 import { readFileSync } from "node:fs";
 import { describe, expect, it } from "vitest";
 import { runCli } from "@/cli";
-import { createWritable } from "../helpers/streams";
+import { createWritable } from "../tests/helpers/streams";
 
 const packageJson = JSON.parse(
-	readFileSync(new URL("../../package.json", import.meta.url), "utf8"),
+	readFileSync(new URL("../package.json", import.meta.url), "utf8"),
 ) as { version: string };
 
 describe("cli", () => {
